+++ packages/sit-onyx/.storybook/vModel.ts
@@ -1,8 +1,15 @@
 import { withVModelDecorator } from "@sit-onyx/storybook-utils";
 
+/**
+ * Custom `withVModelDecorator` filter function, because we want don't want to override values that are managed.
+ */
 export const withOnyxVModelDecorator = withVModelDecorator({
-  filter: ({ table, name }) =>
-    table?.category === "events" &&
-    name.startsWith("update:") &&
-    table.defaultValue?.summary !== "MANAGED_SYMBOL",
+  filter: ({ table, name }, _, array) => {
+    if (table?.category !== "events" || !name.startsWith("update:")) return false;
+    const propName = name.replace(/^update:/, "");
+    return array.every(
+      (argType) =>
+        argType.name !== propName && argType.table?.defaultValue?.summary !== "MANAGED_SYMBOL",
+    );
+  },
 });