+++ packages/sit-onyx/src/components/OnyxSwitch/OnyxSwitch.ct.tsx
@@ -80,7 +80,6 @@ test.describe("Screenshot tests", () => {
 
       // wait for the tooltip to show up reliably
       if (["focus-visible", "hover"].includes(row)) {
-        // eslint-disable-next-line playwright/no-standalone-expect
         await expect(
           component.getByRole("tooltip"),
           `should show error tooltip for ${row} and ${column}`,