+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-beta.57
+
+### Patch Changes
+
+- 78a1019: fix(OnyxNavButton): prop `mobileChildrenOpen` not working
+
+## 1.0.0-beta.56
+
+### Patch Changes
+
+- b6e0b67: fix: Fix uncaught TypeError for components with symbol props
+
 ## 1.0.0-beta.55
 
 ### Major Changes